refactor: modularize build job routing and mapping logic into separate files
